url and form tests are added to each app
